Standardize test layout (use tests directory more)
