+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "coverage-gutters.coverageBaseDir": "target",
-    "rust-analyzer.cargo.features": ["serde", "serde_test"],
+    "rust-analyzer.cargo.features": ["serde"],
     "[rust]": {
         "editor.formatOnSave": true,
         "editor.defaultFormatter": "rust-lang.rust-analyzer"
