+++ Cargo.toml
@@ -12,7 +12,6 @@ repository = "https://github.com/DataDog/libddwaf-rust"
 [dependencies]
 libc = "0.2"
 serde = { version = "1.0", optional = true }
-serde_json = { version = "1.0", optional = true }
 
 [build-dependencies]
 bindgen = "0.72.0"
@@ -20,10 +19,12 @@ flate2 = "1.1"
 reqwest = { version = "0.12", features = ["blocking"] }
 tar = "0.4"
 
+[dev-dependencies]
+serde_json = "1.0"
+
 [features]
 default = ["serde"]
 serde = ["dep:serde"]
-serde_test = ["serde", "dep:serde", "dep:serde_json"]
 static = []
 
 [lints.rust]
